dry up contributor creation for email
